requirements.txt
streamlit
numpy
matplotlib
networkx
scikit-learn
scipy
Pillow
torch
torchvision
